Update Config Data
Changed private data of OPC UA Client Configuration.lvclass so that
there are not seperate items for absolute and relative paths. There was
no need to have seperate items in addition to having a boolean
indicating whether they were absolute or relative paths. This might
change in the future if users find it useful to have some relative and
some absolute paths to server certificates.

All unit tests are passing.