Removed unit tests from library and added error file
UTF tests were moved from inside of the test lvlib to their own folder.
The idea was to be able to move these tests between a Windows and RT
target but this would not work in an auto-build process. All Vis look
changed because moving between computer and crio project items forced a
recompile.